+++ skeptic/dev-resources/skeptic/cljs_fixtures/broken_require.cljs
@@ -1,0 +1,4 @@
+(ns skeptic.cljs-fixtures.broken-require
+  (:require [no.such.namespace-zzz :as nope]))
+
+(defn use-it [] (nope/f 1))